+++ CHANGELOG.rst
@@ -10,11 +10,20 @@ Changelog
 0.4.0 (2024-06-**)
 ------------------
 
-* Renamed :meth:`metalearners.explainer.Explainer.get_feature_importances`. Note this is
+* Renamed :meth:`metalearners.explainer.Explainer.feature_importances`. Note this is
   a breaking change.
 
-* Renamed :meth:`metalearners.metalearner.MetaLearner.get_feature_importances`. Note this
+* Renamed :meth:`metalearners.metalearner.MetaLearner.feature_importances`. Note this
   is a breaking change.
+
+* Renamed :meth:`metalearners.explainer.Explainer.shap_values`. Note this is
+  a breaking change.
+
+* Renamed :meth:`metalearners.metalearner.MetaLearner.shap_values`. Note this
+  is a breaking change.
+
+* Renamed :meth:`metalearners.metalearner.MetaLearner.explainer`. Note this is
+  a breaking change.
 
 
 0.3.0 (2024-06-03)